ball reflect from walls, ballmove, gravity applayed, squere gizmo reflect ball

+++ src/newGizmo/model/Absorber.java
@@ -20,7 +20,7 @@ public class Absorber extends AbstractGizmoModel {
 	@Override
 	public Graphics paint(Graphics g) {
 		g.setColor(Color.BLACK);
-		g.fillRect(0, y, 700, 20);
+		g.fillRect(0, (int)y, 700, 20);
 		return g;
 	}
